requirements.txt
flask
gunicorn
numpy==1.18.3
pillow
tinydb
opencv-python
matplotlib
https://github.com/google-coral/pycoral/releases/download/release-frogfish/tflite_runtime-2.5.0-cp37-cp37m-linux_x86_64.whl